+++ CSC211GroupProject/CSC211GroupProject.vcxproj
@@ -128,6 +128,10 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="GroupProject.cpp" />
+    <ClCompile Include="lineType.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="lineType.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
