+++ CSC211GroupProject/x64/Debug/CSC211Gr.8b5c111b.tlog/CSC211GroupProject.lastbuildstate
@@ -1,2 +1,2 @@
 PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.42.34433:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|C:\Users\Juan\source\repos\CSC211GroupProject\|
+Debug|x64|C:\Users\Juan\source\repos\211GroupProject\|
