+++ CSC211GroupProject/x64/Debug/CSC211Gr.8b5c111b.tlog/Cl.items.tlog
@@ -1,1 +1,2 @@
-C:\Users\Juan\source\repos\CSC211GroupProject\CSC211GroupProject\GroupProject.cpp;C:\Users\Juan\source\repos\CSC211GroupProject\CSC211GroupProject\x64\Debug\GroupProject.obj
+C:\Users\Juan\source\repos\211GroupProject\CSC211GroupProject\GroupProject.cpp;C:\Users\Juan\source\repos\211GroupProject\CSC211GroupProject\x64\Debug\GroupProject.obj
+C:\Users\Juan\source\repos\211GroupProject\CSC211GroupProject\lineType.cpp;C:\Users\Juan\source\repos\211GroupProject\CSC211GroupProject\x64\Debug\lineType.obj
