+++ CSC211GroupProject/x64/Debug/CSC211Gr.8b5c111b.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^C:\USERS\JUAN\SOURCE\REPOS\CSC211GROUPPROJECT\CSC211GROUPPROJECT\X64\DEBUG\GROUPPROJECT.OBJ
-C:\Users\Juan\source\repos\CSC211GroupProject\CSC211GroupProject\x64\Debug\CSC211GroupProject.ilk
+^C:\USERS\JUAN\SOURCE\REPOS\211GROUPPROJECT\CSC211GROUPPROJECT\X64\DEBUG\GROUPPROJECT.OBJ|C:\USERS\JUAN\SOURCE\REPOS\211GROUPPROJECT\CSC211GROUPPROJECT\X64\DEBUG\LINETYPE.OBJ
+C:\Users\Juan\source\repos\211GroupProject\CSC211GroupProject\x64\Debug\CSC211GroupProject.ilk
