+++ CSC211GroupProject/x64/Debug/CSC211GroupProject.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>C:\Users\Juan\source\repos\CSC211GroupProject\x64\Debug\CSC211GroupProject.exe</FullPath>
+      <FullPath>C:\Users\Juan\source\repos\211GroupProject\x64\Debug\CSC211GroupProject.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
