+++ CSC211GroupProject/x64/Debug/CSC211GroupProject.log
@@ -1,2 +1,4 @@
 ﻿  GroupProject.cpp
-  CSC211GroupProject.vcxproj -> C:\Users\Juan\source\repos\CSC211GroupProject\x64\Debug\CSC211GroupProject.exe
+  lineType.cpp
+  Generating Code...
+  CSC211GroupProject.vcxproj -> C:\Users\Juan\source\repos\211GroupProject\x64\Debug\CSC211GroupProject.exe
